Feature: add full response typing for status code, data and headers.
This commit adds additional type information to the route's response
object that allows types to be created that express the combination
of status code, data and headers for each response.

This type information can be used to "narrow" a response type based on
an HTTP status code. This narrowing is useful since it can prevent type
mismatches where a status code is presumed to have one type but in actuality
it returns another.

Additionally, response headers are now associated with data types and status
codes rather than being free form.

A caveat right now is that only there are no response codes like 3XX or
5XX, every code must be explicitly listed.